refactor spaces
