+++ src/aioway/spaces/__init__.py
@@ -6,5 +6,3 @@ from .dims import *
 from .gym import *
 from .media import *
 from .spaces import *
-from .tdicts import *
-from .tensors import *
