+++ src/aioway/spaces/media.py
@@ -10,8 +10,7 @@ from aioway.errors import re_raise_func
 
 from .attrs import Attr
 from .casts import CastedSpaceModule, register_cast
-from .spaces import space_dcls
-from .tensors import TensorSpace
+from .spaces import TensorSpace, space_dcls
 
 __all__ = ["ImageSpace", "ByteImageSpace", "FloatImageSpace"]
