+++ src/aioway/spaces/spaces/__init__.py
@@ -1,0 +1,7 @@
+# Copyright (c) AIoWay Authors - All Rights Reserved
+
+"The core `Space` abstractions."
+
+from .spaces import *
+from .tdicts import *
+from .tensors import *
