+++ src/aioway/spaces/spaces/tdicts.py
