+++ src/aioway/spaces/spaces/tensors.py
@@ -8,12 +8,9 @@ import typing
 import torch
 
 from aioway._api import public_api
-from aioway._torch import is_fake_tensor
+from aioway._torch import Attr, is_fake_tensor
 
 from .spaces import Space, space_dcls
-
-if typing.TYPE_CHECKING:
-    from .attrs import Attr
 
 __all__ = ["TensorSpace"]
 
@@ -26,8 +23,6 @@ class TensorSpace(Space[torch.Tensor], abc.ABC):
     @typing.override
     @typing.final
     def contains(self, tensor: torch.Tensor, /) -> bool:
-        from .attrs import Attr
-
         if not isinstance(tensor, torch.Tensor):
             raise TypeError(f"{type(tensor)} is not a `torch.Tensor`.")
